fix: timeout exceptions for large image messages
The socket raised timeout errors when trying to receive a stream of IMAGE
messages which have a higher resolution (1440x1080). Increasing the socket timeout
lead to problems when trying to send data back (higher latency or never reached the
sending part as always data arrived).
Fixed by adding a helper method which tries to receive exactly the number of bytes
which are expected in a given deadline.
In order to be able to send data back to the server, at max 3 messages will be received
until data can be send back.